requirements.txt
super_gradients
pynndescent==0.5.11
numba==0.58.1
numpy==1.26.4
faiss-cpu
pyvirtualdisplay
Pillow
pyrender
trimesh
smplx
opencv-python
umap-learn
pycocotools
torch==2.3.1
torchvision==0.18.1
gradio
opencv-python
gin-config